nfl_ingestor/requirements.txt
pika==1.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7